Move Cas2v2 App Status finder
This was previously in a reporting package, which is incorrect.

This commit also removes two unused functions on the `Cas2v2PersistedApplicationStatusFinder`
